rsbuild: watch logic
