+++ packages/rspeedy/core/src/cli/dev.ts
@@ -12,6 +12,7 @@ import type { CommonOptions } from './commands.js'
 import { createRspeedy } from '../create-rspeedy.js'
 import { exit } from './exit.js'
 import { init } from './init.js'
+import { getWatchedFiles, watchFiles } from './watch.js'
 
 export interface DevOptions extends CommonOptions {
   base?: string | undefined
@@ -30,18 +31,7 @@ export async function dev(
       devOptions,
     )
 
-    const watchedFiles = [configPath]
-
-    if (Array.isArray(rspeedyConfig.dev?.watchFiles)) {
-      watchedFiles.push(
-        ...rspeedyConfig.dev.watchFiles
-          .filter(item => item.type === 'reload-server')
-          .flatMap(item => item.paths),
-      )
-    } else if (rspeedyConfig.dev?.watchFiles?.type === 'reload-server') {
-      const { paths } = rspeedyConfig.dev.watchFiles
-      watchedFiles.push(...Array.isArray(paths) ? paths : [paths])
-    }
+    const watchedFiles = getWatchedFiles(configPath, rspeedyConfig)
 
     await watchFiles(
       watchedFiles.map(filePath =>
@@ -67,53 +57,5 @@ export async function dev(
   } catch (error) {
     logger.error(error)
     exit(1)
-  }
-}
-
-async function watchFiles(
-  files: string[],
-  callback: (
-    filePath: string,
-    startTime: number,
-    event: string,
-  ) => Promise<void>,
-) {
-  const chokidar = await import('chokidar')
-  const watcher = chokidar.default.watch(files, {
-    // do not trigger add for initial files
-    ignoreInitial: true,
-    // If watching fails due to read permissions, the errors will be suppressed silently.
-    ignorePermissionErrors: true,
-  })
-
-  const cb = debounce(
-    (event: string, filePath: string) => {
-      const startTime = Date.now()
-      void watcher.close().then(() => callback(filePath, startTime, event))
-    },
-    // set 300ms debounce to avoid restart frequently
-    300,
-  )
-
-  watcher.once('add', cb.bind(null, 'add'))
-  watcher.once('change', cb.bind(null, 'change'))
-  watcher.once('unlink', cb.bind(null, 'unlink'))
-}
-
-// biome-ignore lint/suspicious/noExplicitAny: Make TS happy
-function debounce<T extends (...args: any[]) => void>(
-  func: T,
-  wait: number,
-): (...args: Parameters<T>) => void {
-  let timeoutId: ReturnType<typeof setTimeout> | null = null
-
-  return (...args: Parameters<T>) => {
-    if (timeoutId !== null) {
-      clearTimeout(timeoutId)
-    }
-
-    timeoutId = setTimeout(() => {
-      func(...args)
-    }, wait)
   }
 }
